% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countgiveups.R
\name{countGiveUps}
\alias{countGiveUps}
\title{countGiveUps}
\usage{
countGiveUps(X, numAtts)
}
\arguments{
\item{X, }{a response matrix of dimension nS x nI}

\item{numAtts, }{a vector of the max attempts allowed per item}
}
\value{
a list containing the counts of giveups and the counts of total wrong attempts (each vector of length nS)
}
\description{
countGiveUps
}

