% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multattfuncs.R
\name{generate.data}
\alias{generate.data}
\title{Generate response matrix given model and parameters}
\usage{
generate.data(model, pparams, iparams, scoreSeed)
}
\arguments{
\item{model}{One of "GRM","GPCM", "SM"}

\item{pparams}{A list containing at least theta vector}

\item{iparams}{A list containing discrimination and threshold parameters (a,b)}

\item{scoreSeed}{A random seed}
}
\value{
A simulated response matrix
}
\description{
Generate response matrix given model and parameters
}
\keyword{internal}

