% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefuncs.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse Logit, aka Logistic function}
\usage{
inv.logit(z)
}
\arguments{
\item{z}{a real valued number.}
}
\value{
Output is the value of the function 1/(1+exp(-z)), on the interval [0,1]
}
\description{
\code{inv.logit} computes 1/(1+exp(-z)), the inverse of
\code{\link{logit}}
}
\examples{
z <- 5
inv.logit(z)
logit(inv.logit(z))
}

