% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefuncs.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(z)
}
\arguments{
\item{z}{a real value in the open interval (0,1), otherwise NaNs produced.}
}
\value{
Output is the value of the function log(z/(1-z))
}
\description{
\code{logit} computes log(z/(1-z)), whereas
\code{\link{inv.logit}} return the inverse
}
\examples{
z <- 5
logit(z)
inv.logit(logit(z))
}

