% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mima.R
\name{makeXminus}
\alias{makeXminus}
\title{makeXminus, generate reduced (Xminus) matrix from expanded (Xplus) response data
In Xminus, all but the last incorrect attempt and/or first correct attempt are unobserved}
\usage{
makeXminus(responses, attempts)
}
\arguments{
\item{responses, }{Xplus matrix}

\item{attempts, }{a vector of length = numItems, maximum attempts allowed}
}
\value{
Xplus matrix
}
\description{
makeXminus, generate reduced (Xminus) matrix from expanded (Xplus) response data
In Xminus, all but the last incorrect attempt and/or first correct attempt are unobserved
}

