% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mima.R
\name{makeXnaught}
\alias{makeXnaught}
\title{makeXnaught, generate standard sequential matrix from expanded (Xplus) response data
In Xnaught, all attempts after correct are unobserved}
\usage{
makeXnaught(responses, attempts)
}
\arguments{
\item{responses, }{Xplus matrix}

\item{attempts, }{a vector of length = numItems, maximum attempts allowed}
}
\value{
Xnaught matrix
}
\description{
makeXnaught, generate standard sequential matrix from expanded (Xplus) response data
In Xnaught, all attempts after correct are unobserved
}

