% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mima.R
\name{makeXplus}
\alias{makeXplus}
\title{makeXplus, generate expanded matrix from observed attempt count (OAC) response data.
In Xplus, all attempts after correct are imputed to be observed correct}
\usage{
makeXplus(U, allowedAttempts)
}
\arguments{
\item{U, }{response matrix in OAC format}

\item{allowedAttempts, }{a vector of length ncol(U) with maximum attempts allowed}
}
\value{
Xplus matrix with ncol = sum(allowedAttempts)
}
\description{
makeXplus, generate expanded matrix from observed attempt count (OAC) response data.
In Xplus, all attempts after correct are imputed to be observed correct
}

