% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countgiveups.R
\name{maxOutOAC}
\alias{maxOutOAC}
\title{maxOutOAC}
\usage{
maxOutOAC(X, numAtts)
}
\arguments{
\item{numAtts}{}
}
\value{
a matrix with OAC values less than or equal to max attempts
}
\description{
maxOutOAC
}

