% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianattempts.R
\name{medianattempts}
\alias{medianattempts}
\title{Median Attempts}
\usage{
medianattempts(U, maxAttempts = 99, useAbs = F)
}
\arguments{
\item{U}{a response matrix containing observed attempts to correct. U is
integer-valued, and $U_ij$ = n, where n <0 if the nth attempt was (still)
incorrect.}

\item{maxAttempts}{an integer denoting the cutoff M, such that attempt values
greater than M are ignored in finding the median. The only reason to use
this argument is to remediate abuse/gaming of the system.}

\item{useAbs}{Boolean, default is \code{FALSE}, in which case only correct
attempts are included in finding the median. If \code{TRUE} then incorrect
attempts are also counted after taking absolute values.}
}
\value{
a vector of length \code{nrow(U)}
}
\description{
This function is useful in response matrices with missing data (NAs) and also
possible gaming behaviors, such as students with very large numbers of observed attempts
}
\examples{

medianattempts(responses)
}

