% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multattfuncs.R
\name{multatt.estim}
\alias{multatt.estim}
\title{Estimate parameters given multiple attempt data and model}
\usage{
multatt.estim(data, datatype, numAtts, model)
}
\arguments{
\item{data}{A response matrix or data frame, with ncols = number of items
(nI)}

\item{datatype}{Either "graded" or "attempts". Graded type means category
labels run from 0...M where M is the highest category and corresponds to
answering correct on the first attempt. A value of 0 corresponds to
incorrectly exhausting all allowable attempts. In the "attempt"
representation, values range from -M to M, excluding 0. See
\code{\link{recodeGRM2SM}} and \code{\link{recodeSM2GRM}} for interchanging
between these representations, although the mapping is not exactly
one-to-one.}

\item{numAtts}{a vector of length nI containing max allowed attempt numbers
for each item}

\item{model}{One of "GRM", "GPCM", "SM.Rasch", "SM.2pj", "SM.2pjx",
"MIMA.Rasch", "MIMA.2pj", "MIMA.2pjx"}
}
\value{
A list containing the fitobject (fitObj), item parameters (itemPars),
  and person parameters (personPars)
}
\description{
Estimate parameters given multiple attempt data and model
}
\examples{
nS = 500
nI = 9
numAtts <- c(2,2,2,3,3,3,4,4,4)
randomSeeds <- c(21,22)
simGRM <- multatt.sim(model="GRM", nS, nI, numAtts, randomSeeds)
simSMR <- multatt.sim(model="SM.Rasch", nS, nI, numAtts, randomSeeds)
#estGRMwithSMR <- multatt.estim(simGRM$X, datatype="graded", numAtts, model="SM.Rasch")
#estSMRwithGRM <- multatt.estim(simSMR$X, datatype="attempts", numAtts, model="GRM")
}

