% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multattfuncs.R
\name{multatt.sim}
\alias{multatt.sim}
\title{Simulate multiple attempt data}
\usage{
multatt.sim(model, nS, nI, numAtts, randomSeeds, orderedCats = FALSE,
  pParamType = "nogiveup", paramcontrol = list(mu.theta = 0, sd.theta = 1,
  multimu = c(0, -3), Sigma = matrix(c(1, -0.45, -0.45, 1), nrow = 2), mu.a = 0,
  sd.mu.a = 0.5, sd.a = 0.2, mu.b = 0, sd.mu.b = 1, sd.b = 0.3))
}
\arguments{
\item{model}{One of "GRM","GPCM","SM.Rasch","SM.2pj","SM.2pjx"}

\item{nS}{number of examinees}

\item{nI}{number of items}

\item{numAtts}{a vector of length nI containing allowed attempt numbers for each item}

\item{randomSeeds}{Two values, one for input parameters and one for the generated responses}

\item{orderedCats}{Logical; applies to the sequential model}

\item{paramcontrol}{A list of distribution parameters for threshold and discrimination parameters}
}
\value{
Output is a list containing pparams (person parameters), iparams
  (item parameters), and X (response matrix) integer.
}
\description{
\code{multatt.sim} returns the simulated values for person parameters,
item parameters, and a polytomous response matrix. The response matrix will
either follow the "graded" response convention if the generating model is GRM
or GPCM or the "attempts" count convention, if a sequential model is used.
}
\examples{
nS = 10
nI = 5
numAtts <- c(2,2,3,3,3)
randomSeeds <- c(2015,2016)
simGRM <- multatt.sim(model="GRM", nS, nI, numAtts, randomSeeds)
simSMR <- multatt.sim(model="SM.Rasch", nS, nI, numAtts, randomSeeds)

}

