% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{reCodeGRM2SM}
\alias{reCodeGRM2SM}
\title{reCode GRM->SM}
\usage{
reCodeGRM2SM(U, itemAllowedAttempts)
}
\arguments{
\item{U}{a response matrix in graded polytomous format}

\item{itemAllowedAttempts}{a vector of length numItems, where each element is the maximum allowed attempts.}
}
\value{
Output is a response matrix in OAC form, see \code{\link{responses}}
}
\description{
This function and \code{\link{reCodeSM2GRM}} convert data representations between from
OAC (observed attempt counts) and graded polytomous responses. This mapping is NOT one-to-one.
}
\examples{
tmp <- responses[1:10,1:5]
tmp2 <- reCodeSM2GRM(tmp, rep(3,ncol(tmp)))
tmp3 <- reCodeGRM2SM(tmp2, rep(3,ncol(tmp)))
}

