% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{reCodeSM2GRM}
\alias{reCodeSM2GRM}
\title{reCode SM->GRM}
\usage{
reCodeSM2GRM(U, itemAllowedAttempts)
}
\arguments{
\item{U}{a response matrix in OAC format, see \code{\link{responses}}}

\item{itemAllowedAttempts}{a vector of length numItems, where each element is the maximum allowed attempts
for converting to an graded polytomous representation.}
}
\value{
Output is a response matrix in ordered form.
}
\description{
This function and \code{\link{reCodeGRM2SM}} convert data representations between OAC (observed attempt counts)
and graded polytomous scores. This mapping is NOT one-to-one.
}
\examples{
tmp <- responses[1:10,1:5]
reCodeSM2GRM(tmp, rep(3,ncol(tmp)))

}

