% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sim.R
\name{simulate.a.sm}
\alias{simulate.a.sm}
\title{Title}
\usage{
\method{simulate}{a.sm}(n.item, n.atts, mu.a = 0, sd.mu.a = 0.5,
  sd.a = 0.2, seed)
}
\arguments{
\item{seed}{}
}
\value{
a parameters
}
\description{
Title
}

