% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sim.R
\name{simulate.b.sm}
\alias{simulate.b.sm}
\title{Title}
\usage{
\method{simulate}{b.sm}(n.item, n.atts, orderedCats = FALSE, mu.b = 0,
  sd.mu.b = 1, sd.b = 0.3, seed)
}
\arguments{
\item{seed}{}
}
\value{
b parameters
}
\description{
Title
}

