% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sim.R
\name{simulate.multidim}
\alias{simulate.multidim}
\title{Title}
\usage{
\method{simulate}{multidim}(n.person, type = "nogiveup", mu = c(0, -3),
  Sigma = matrix(c(1, -0.45, -0.45, 1), nrow = 2), seed)
}
\arguments{
\item{n.person}{number of examinees}

\item{type}{in c("nogiveup","indep","custom") where "nogiveup" sets all giveupitudes to 0, indep samples theta and
giveupitude independently, and custom samples from a multivariate normal distribution followed by an exponential transformation
for giveupitude.}

\item{mu}{means vector for theta and giveupitude}

\item{Sigma}{covariance matrix}

\item{seed}{random seed}
}
\value{
pparams, a list containing the theta and giveupitude vectors
}
\description{
Title
}

