% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsm.R
\name{simulate.sm}
\alias{simulate.sm}
\title{Simulate Sequential Model}
\usage{
\method{simulate}{sm}(a, b, theta, omittitude = rep(0, length(theta)),
  giveupitude = rep(0, length(theta)), seed, debug = F)
}
\arguments{
\item{a}{discrimination parameters; a list of length numItems, where each element is a vector}

\item{b}{threshold parameters; a list of length numItems, where each element is a vector}

\item{theta}{student skills; a vector of length numStudents}

\item{omittitude}{a vector of length numStudents encoding the probability of omitting an item entirely (no attempts made)}

\item{giveupitude}{a vector of length numStudents encoding the probability of not attempting further after an incorrect attempt}

\item{seed}{random seed}

\item{debug}{BOOLEAN flag}
}
\value{
Output is a response matrix of integers, see sample data set \code{\link{responses}}
}
\description{
An internal function used by \code{\link{multatt.sim}}
}

